requirements.txt
python-fasthtml
python-dotenv
cohere
reportlab
lancedb